Move log out handling in the home activity
