+++ app/src/main/java/ch/epfl/qedit/view/home/HomeActivity.java
@@ -1,23 +1,31 @@
 package ch.epfl.qedit.view.home;
 
-import static ch.epfl.qedit.view.login.Util.USER;
-
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.MenuItem;
 import android.widget.Toast;
+
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.ActionBarDrawerToggle;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.view.GravityCompat;
 import androidx.drawerlayout.widget.DrawerLayout;
+
+import com.google.android.material.navigation.NavigationView;
+import com.google.firebase.auth.FirebaseAuth;
+
+import java.util.Objects;
+
 import ch.epfl.qedit.R;
 import ch.epfl.qedit.model.User;
 import ch.epfl.qedit.util.LocaleHelper;
 import ch.epfl.qedit.view.QR.ScannerActivity;
-import com.google.android.material.navigation.NavigationView;
-import java.util.Objects;
+import ch.epfl.qedit.view.login.LogInActivity;
+import ch.epfl.qedit.view.login.Util;
+
+import static ch.epfl.qedit.view.login.Util.USER;
+import static ch.epfl.qedit.view.login.Util.USER_ID;
 
 public class HomeActivity extends AppCompatActivity
         implements NavigationView.OnNavigationItemSelectedListener {
@@ -112,5 +120,37 @@ public class HomeActivity extends AppCompatActivity
                         R.string.navigation_drawer_close);
         drawer.addDrawerListener(toggle);
         toggle.syncState();
+    }
+
+    // This method will be called when an item of the top bar is clicked on
+    @Override
+    public boolean onOptionsItemSelected(@NonNull MenuItem item) {
+        if (item.getItemId() == R.id.log_out) {
+            logOut();
+            return true; // the event was handled
+        }
+
+        return false; // the event was not handled
+    }
+
+    private void logOut() {
+        // Retrieve cached user id
+        String uid = Util.getStringInPrefs(this, USER_ID);
+
+        // Remove cached user id
+        Util.removeStringInPrefs(this, USER_ID);
+
+        // Log out
+        FirebaseAuth.getInstance().signOut();
+
+        // Go to log in activity
+        startActivity(new Intent(this, LogInActivity.class));
+
+        // Show toast
+        Toast.makeText(
+                this,
+                getResources().getString(R.string.log_out_success),
+                Toast.LENGTH_SHORT)
+                .show();
     }
 }
